Improve scrolling when holding h/l/j/k down
When you hold an arrow key down, the scrolling is sped up. With this
commit, the `h`, `l`, `j` and `k` do too, to be as close the the arrow
keys as possible.

Note that this includes a minor breaking API change: The object passed
to modes' `onInput`, and thus the object passed to commands, no longer
has the `uiEvent` property. Instead, there's an `event` property, which
_always_ is an `Event`. `vim.isUIEvent(event)` can be used to check if
it is a UI event or not.