Only apply `vimfx.addKeyOverrides` in Normal mode
Previously, the key overrides were applied in _all_ modes, and the
current mode name was passed as the second argument to the matching
functions. However, it was way to easy to forget to also check the mode
in the matching functions (even the documentation forgot to!). For
example, if you added `j` as an override for some page, you then
couldn't press `j` in Hints mode to match a hint marker.

This commit fixes this problem by simply _only_ applying the key
overrides in Normal mode. I can't think of any use cases for wanting key
overrides in any other mode.

This is technically a minor breaking API change.

Fixes #812.