Fix selection clipboard handling
This is only for systems with a "selection clipboard."

1. Copy text using one of the following methods:

   - `yv<hint>`
   - `av<hint>y`
   - `yf<hint>` on a contenteditable element

2. Use the `p` or `P` command.

Unlike other commands that copy text to the clipboard, such as `yy` and
`v<hint>wy`, the above methods previously didn't affect the selection
clipboard. This caused the `p` and `P` commands to operate on some other
text than expected.

This commit makes sure that all commands that copy text to the clipboard
also affect the selection clipboard (in a somewhat hacky, but working,
way).

Fixes #818.