Fix z-index for hint markers
The hint markers are sorted by element area before assigning their
`z-index`es. This is so that the hint markers of larger elements appear
on top of hint markers for smaller elements. However,
`marker.wrapper.area` (which is `undefined`) was accidentally used
instead of `marker.wrapper.shape.area`. This caused the sorting to
essentially do nothing (since Firefox's `.sort` is stable).