Fix `<escape>` handling in Normal mode
When pressing `<escape>` inside a text input, `<escape>` is not supposed
to be passed on to the page (like is otherwise done), to allow blurring
text inputs inside modals without also closing the modal. There already
was logic for this, but it didn't work because the check for the
currently selected element and the running of the `<escape>` command was
run in the wrong order.