Improve full page scrolling adjustment on some pages
On medium.com, there is a fixed footer. However, it is not
`position: fixed;` like fixed element usually are. Instead, it's
`position: absolute;` (and the page scroll is inside an adjacent element
instead of on `<html>`). This commit allows elements with
`position: absolute;` to as headers and footers as well.