Improve markable element area comparisons
Boxes that appear to be exactly the same size should be given "equally"
good hints, favoring earlier ones when needed. Sometimes, though, their
areas differ from a fraction of a pixel up to a couple of pixels.

This commits changes how the comparisons of marker weights (which
effectively are element areas), so that elements whose area differ by
less than 10px are considered to have the same area. That results in the
expected hints for same-sized boxes.

Refs. #831: The browser tabs often had this problem. At first, all tabs
could be given the expected hints: The first one getting the best one
etc. Opening a new tab, the best hint could suddenly be given to some
random tab in the middle.