Find `<label>` elements in the first pass for `f`
It is common to implement custom styled checkboxes and radiobuttons by
hiding the actual `<input>` element and style its `<label>` instead.
Therefore it makes sense to find them in the first pass, so you don't
have to wait for their labels so long. This does not seem to degrade the
performance of the first pass.