Blur the location bar consistently
Refs. #831: Let's say you added the custom command that focuses the
location bar, sets its text to '% ' and opens the autocomplete popup.
You set the shortcut to `q`. If you then press `q<escape>`, you'll end
up with '% ' being left in the location bar. Usually, the location bar
is cleared when blurring it. This is because of that the autocomplete
popup is open. This commit makes sure to close the location bar
autocomplete popup before blurring, fixing the problem.