Fix tests in Firefox Nightly
Firefox have either moved the Assert.jsm module, or stopped exposing it.
I'm tired of that happening (it's not the first time), so I decided to
implement the small part of Assert.jsm that was actually used in VimFx's
test on my own. On the upside, this resulted in nicer error messages
when tests fail.