Scroll closest scrollable parent
If you click a tweet on Twitter, a modal with the tweet as well as
responses to it pop up. That modal can be scrollable if there are enough
responses. The container for the main tweet is focused. It feels very
natural that the _modal_ should scroll when using VimFx commands, not
the page behind it. That used not to be the case, but now is. This was
implemented by not just checking if the active element _itself_ is
scrollable, but also whether one of its _parents_ are scrollable, and
using the closest scrollable parent (if any).