Fix TypeError when hitting ESC
When hitting esc a TypeError was shown in the browser console
(Ctrl-Shift-J), except when an element is fullscreened. The location of
the error (i.e. commands-frame.js:621) is only shown on special pages
(like about:* or chrome://), making the problem harder to locate.
see also: functionality introduced in commit ab17744
> TypeError: The expression cannot be converted to return the specified type. commands-frame.js:621