replace custom status panel with default one
In Mozilla's efforts to get rid of XBL[1] they changed the way the
status panel works. Up to Firefox 68, the changes meant that the panel
just didn't show up, but with Firefox 69 and onwards, it broke the
display of the browser. This change replaces the custom (secondary)
status bar and uses the main one for displaying infos and errors.
See also: bug[2] and changeset[3] of the removal of <statusbar>

[1]: http://arewexblstill.com/
[2]: https://bugzilla.mozilla.org/show_bug.cgi?id=1444891
[3]: https://hg.mozilla.org/mozilla-central/rev/18bbbdc02213