fix hints mode and help popups
entering any of these modes adds an overlay above the <browser> element.
In mozilla69 the style property 'display: -moz-box' was removed from '*'
in minimal-xul.css. When the overlay was displayed without this
property, the <browser> would be resized to 25x25px to make room for the
overlay instead of displaying them on top of each other.

keep an eye on http://bugzil.la/939428