fix hints mode on about:addons and about:preferences in nightly 70
In Bug 1550801, all XUL documents were loaded as HTMLDocuments, which
caused the isXULDocument test to fail. The new test is modeled after a
snippet from devtools/.

This change caused hints mode to break in about:preferences and
about:addons, since (some of?) the de-XBL'd custom elements' <label>s
return the id (as string) when reading theLabelElement.control on a
label instead of an HTMLElement. VimFx passes this value to a function
that assumes the value has a .getClientRects() method.

Only these two pages were affected, since these are the only in-content
pages (other than about:downloads) in Firefox 70 that still have a .xul
document at the root.