fix 'eb' for fx71
browser.xhtml was restructured between firefox 69 and 71, and hints were
not displaying any more (they still worked, adding to the confusion). It
turns out that moving the hints-containert before #navigator-toolbox is
sufficient. But insertBefore()ing that element breaks Waterfox 56, as
this element does not exist there (which is a bit strange, as the docs
say, insertBefore should degrade gracefully to appendChild in such a
case). However, we can just insert it as the first element below
the #main-window element, which works a treat.