move from deprecated Cu.import to ChromeUtils.import
In most cases, this was a drop-in replacement, but in bootstrap.coffee,
we used the side effect of polluting the global namespace as a handy way
to inject the `Services` object into the user's config.js file.

As a bonus, the codebase has been rid of an inconsistency between
Components.utils.import() and its shorthand Cu.import(), since both are
now gone. We also use the destructuring assignment consistently now.

Both config.js and frame.js were checked to still have the same exposed
API (mainly concerned with Services and console, since that was removed
from bootstrap.coffee).

see also: Bug 1765167