fix installation on localized/l10n versions of firefox
resource://gre/modules/addons/XPIInstall.jsm tries to get the Addon's
name in loadManifest(). On localized versions supported by VimFx, this
is read from install.rdf's <em:localized> section. We assumed that when
this block does not include an <em:name> element, that it would use the
one from the non-localized section. That does not seem to have been
true, ever. Even l10n-Firefox 52 balks here (using 0.21.0 and 0.22.2),
as well as i10n-Firefox 68.

fixes: https://github.com/akhodakivskiy/VimFx/issues/982