handle clicking elements that went hidden during click handler
dispatchDOMEventViaPresShellForTesting cannot handle display:none
elements, so we need to fall back to element.click(). the following
snippet reproduces the bug nicely; activate hints (`f`), then wait for
the element to disappear, then follow the hint:

  <!DOCTYPE html>
  <button id=b onclick="alert(1)">click</button>
  <script>
  window.setTimeout(function(){ b.style.display = "none"; }, 5000);
  </script>

this also happens on items of the main menu on https://gigglehd.com/gg/,
where each menu and each menu item have an onmouseout handler that hides
the menu. this handler is triggered by VimFx's automatically fired
'hover-end' simulated mouse events, which occur immediately before the
'hover-start' and the 'click' events the user requested.

as documented in the comment, our current implementation is not
spec-compliant, but fast and works-for-now. more precisely, the browser
is supposed to return null for position:fixed elements, but firefox
doesn't. should that change, this monstrosity should do the trick:

  if element.offsetParent != null
    return false # definitely visible.
  if element.nodeName in ['BODY', 'HTML']
    return false # special case; assume always visible.

  # element is hidden (likely) or directly has position:fixed (unlikely)
  style = element.ownerGlobal.getComputedStyle(element)
  if style.position == 'fixed'
    if style.display == 'none'
      return true
    else
      # it's a fixed element, but not directly hidden. since display
      # is not inherited, we must check parent element manually.
      return isDisplayNone(element.parentNode)
  else
    # it's not position:fixed, hence it really is hidden.
    return true